+++ api/index.js
@@ -5,12 +5,8 @@ const cors = require('cors')
 const cookieParser = require('cookie-parser')
 const responseTime = require('response-time')
 const app = express()
-const pkg = require('../package.json')
 
-require('dotenv').config({ path: path.join(process.cwd(), '.env') })
-
-// Default server configuration
-const $config = require('../config')
+require('dotenv').config({ path: path.resolve('.env') })
 
 // Functionality
 const database = require('../database')
@@ -18,8 +14,14 @@ const router = require('./router')
 const controllers = require('./router/controllers')
 const plugins = require('./plugins')
 
-const createServer = async object => {
-  const config = await validateConfig(object).catch(err => err)
+// Services
+const validateConfig = require('./services/validate-config')
+
+const createServer = async (object) => {
+  const config = await validateConfig(object).catch(err => {
+    console.error('[validateConfig]', err)
+    process.exit(1)
+  })
   const HOST = config.host || 'localhost'
   const PORT = config.port || 3000
 
@@ -28,10 +30,22 @@ const createServer = async object => {
   app.use(cookieParser())
   app.use(cors(config.cors))
   app.use(initializer(config)) // Set req.app properties
-  app.use('/api', timer, router(config.api)) // Default plugin is api
-  app.use(`/api/auth`, timer, router(plugins.auth)) // Default plugin for auth
+  app.use(path.join(config.api.base), timer, router(config.api)) // Default plugin is api
+  app.use(path.join(config.api.base, 'auth'), timer, router(plugins.auth)) // Default plugin for auth
 
-  // Set plugins/hydrate plugins
+  // Set static folders
+  for (const static of config.api.static) {
+    // virtual path
+    const Base = path.join(config.api.base, static.base || static.folder.replace('.', '')).replace(/\\/g, '/')
+    // physical path
+    const Path = (static.fullpath || path.join(process.cwd(), static.folder)).replace(/\\/g, '/')
+
+    // set static folder
+    app.use(Base, express.static(Path, static.options))
+    // console.log(Base, Path)
+  }
+
+  // Set plugins
   // if (config.api.plugins) {
   //   // additional routes
   //   for (const name in config.api.plugins) {
@@ -61,27 +75,6 @@ const createServer = async object => {
   return app
 }
 
-const validateConfig = async (config) => {
-  if (config.database && !(config.database.development || config.database.production)) {
-    throw new Error('Invalid database config')
-  }
-
-  if (!config.api) {
-    throw new Error('Invalid api config')
-  }
-
-  if (config.api.routes && typeof config.api.routes !== 'object') {
-    throw new Error('Invalid api routes config')
-  }
-
-  // assign default configs if not provided
-  for (const key in $config) {
-    if (!config[key]) config[key] = $config[key]
-  }
-
-  return config
-}
-
 const timer = responseTime((req, res, time) => {
   console.log(`\x1b[33m[${req.method}]\x1b[0m \x1b[34m${req.url}\x1b[0m ${time.toFixed(0)}ms`);
 })
@@ -92,13 +85,8 @@ const initializer = (config) => (req, res, next) => {
   req.app.controllers = controllers
 
   // app name
-  res.set('X-Powered-By', config.poweredBy || `Blue Server v${pkg.version}`)
+  res.set('X-Powered-By', config.poweredBy)
   next()
-}
-
-// development
-if (process.argv[2] === '--dev') {
-  createServer($config)
 }
 
 module.exports = { createServer }